Se incluye proyecto de Test para depurar el código, así como mejoras en la lógica.

+++ TODSLibreria/SimplexEntity/EcuacionVectorial.cs
@@ -11,6 +11,6 @@ namespace TODSLibreria.SimplexEntity
         public IDictionary<string, double> CuerpoVector { get; set; }
         public IEnumerable<string> NombresVariables { get { return CuerpoVector.Select(r => r.Key); } }
         public IEnumerable<double> CuerpoNum { get { return CuerpoVector.Select(r => r.Value); } }
-        public float TerminoIndependiente { get; set; }
+        public double TerminoIndependiente { get; set; }
     }
 }
